Rename model drilldown to tera-model-drilldown
