+++ packages/client/hmi-client/src/components/workflow/ops/model/mod.ts
@@ -1,6 +1,6 @@
 import { ModelOperation as operation } from './model-operation';
 import node from './tera-model-node.vue';
-import drilldown from './tera-model-workflow-wrapper.vue';
+import drilldown from './tera-model-drilldown.vue';
 
 const name = operation.name;
