+++ packages/client/hmi-client/src/components/workflow/ops/model/tera-model-drilldown.vue
